test(build): split TestMain out of test_generate_agents.py
The previous commit's two new CLI regression tests pushed
tests/test_generate_agents.py from 431 to 584 lines, tripping the
whole-repo taste-lint count ratchet (576 > baseline 575): a genuinely
new file-size violation, not a pre-existing one like the other files
this PR already touches.

TestMain (CLI entry-point tests) is also the most natural split point
on cohesion grounds, not just line count: it exercises main()'s
argument parsing and path resolution, a different concern from
TestGenerateAgents (direct calls into the orchestration function) and
TestReadPlatformConfig (config parsing) that stay in the original
file. Moved to tests/test_generate_agents_main.py, which imports
_create_test_structure from tests.test_generate_agents rather than
duplicating it.

Ratchet back at baseline (575); all 150 tests across both files pass.

Refs #5327
